STILL BUGGY
checking in
the bug has to do with when trigroups are joined.  If the fixed group
has an anchor triangle index greater than the moved group, things fail.
The other direction (fixed anchor triangle index < moved anchor tri idx)
and it works fine.
Still working on the fix in JoinTriangle but wanted to checkpoint.