still major bugs but huge speed increase
taking advantage of pixi.js to speed up rendering
instead of doing pan/zoom calculations by hand.

major bugs still due to misalignment of realized
triangles.  I think any method of placement that
relies on vertex points being exact is doomed to failure.
We should make the convention that the realization
is a 'best effort' approach and any algorithm to
traverse boundaries or the like needs to rely on
meta information. (todo)